Move log loading (fromEntry) tests to their own file
